Changed from bb8 to deadpool...

+++ src/handlers/admin/conn.rs
@@ -2,7 +2,7 @@ use crate::{error::AppResult, repositories::UserRepositoryTrait, states::UserSta
 use axum::{extract::State, Json};
 
 #[inline(always)]
-pub async fn pool_status(State(state): State<UserState>) -> AppResult<Json<(u32, u32)>> {
+pub async fn pool_status(State(state): State<UserState>) -> AppResult<Json<(usize, usize)>> {
     let (redis, pg) = state.user_repository.active_pools();
 
     Ok(Json((redis, pg)))
